+++ src/translations/nl/redactor.php
@@ -1,0 +1,20 @@
+<?php
+return [
+    'Add head' => 'Tabelkop invoegen',
+    'Delete column' => 'Kolom verwijderen',
+    'Delete head' => 'Tabelkop verwijderen',
+    'Delete row' => 'Rij verwijderen',
+    'Delete table' => 'Tabel verwijderen',
+    'Fullscreen' => 'Volledig scherm',
+    'Insert column left' => 'Kolom links invoegen',
+    'Insert column right' => 'Kolom rechts invoegen',
+    'Insert Page Break' => 'Pagina-einde invoegen',
+    'Insert row above' => 'Rij hierboven invoegen',
+    'Insert row below' => 'Rij hieronder invoegen',
+    'Insert table' => 'Tabel invoegen',
+    'Link to a category' => 'Link naar een categorie',
+    'Link to an entry' => 'Link naar inhoud',
+    'Table' => 'Tabel',
+    'Video' => 'Video',
+    'Video Embed Code or Youtube/Vimeo Link' => 'Video Embed Code of Youtube/Vimeo Link',
+];